very shitty way of getting the keybind to work

+++ src/Utils.cs
@@ -255,7 +255,7 @@ namespace ProjectApparatus
                 Event e = Event.current;
                 if (e.isKey && e.type == EventType.KeyDown)
                 {
-                    bind.bind = ((int)e.keyCode);
+                    bind.bind = e.keyCode;
                     bind.settingKeybind = false;
                 }
             }
